requirements.txt
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.0
pyserial==2.7
pygame==1.9.2a0
